+++ doc/changelog.rst
@@ -3,7 +3,7 @@ Changelog
 
 v0.2 (*unreleased*)
 -------------------
-- Support the :rst:directive:`testcode`, :rst:directive:`testsetup` and :rst:directive:`testcleanup` directives.
+- Support the :rst:dir:`testcode`, :rst:dir:`testsetup` and :rst:dir:`testcleanup` directives.
 
 
 v0.1.2 (31 August 2020)